+++ src/server/maupod/cmd/mediainfo/main.go
@@ -7,7 +7,7 @@ import (
 	"time"
 
 	"github.com/mauleyzaola/maupod/src/server/pkg/helpers"
-	"github.com/mauleyzaola/maupod/src/server/pkg/rule"
+	"github.com/mauleyzaola/maupod/src/server/pkg/rules"
 	"github.com/mauleyzaola/maupod/src/server/pkg/simplelog"
 	"github.com/mauleyzaola/maupod/src/server/pkg/types"
 	"github.com/spf13/viper"
@@ -36,12 +36,12 @@ func run() error {
 	logger = &simplelog.Log{}
 	logger.Init()
 
-	config, err := rule.ConfigurationParse()
+	config, err := rules.ConfigurationParse()
 	if err != nil {
 		return err
 	}
 
-	if err = rule.ConfigurationValidate(config); err != nil {
+	if err = rules.ConfigurationValidate(config); err != nil {
 		return err
 	}
