+++ src/server/pkg/rules/artwork.go
@@ -1,4 +1,4 @@
-package rule
+package rules
 
 import "github.com/mauleyzaola/maupod/src/server/pkg/pb"
