+++ src/server/pkg/rules/configuration.go
@@ -1,4 +1,4 @@
-package rule
+package rules
 
 import (
 	"errors"
