+++ src/server/pkg/rules/file_store.go
@@ -1,4 +1,4 @@
-package rule
+package rules
 
 import (
 	"errors"
