+++ src/server/pkg/rules/media.go
@@ -1,4 +1,4 @@
-package rule
+package rules
 
 import (
 	"errors"
